fix: survey modal
