+++ book-talk-ts/packages/web/src/hooks/domain/useDebateRealtimeConnection.tsx
@@ -6,6 +6,7 @@ import {
   type MemberInfo,
   type RoundType,
 } from '@src/externals/debate';
+import { createSurvey } from '@src/externals/survey';
 import {
   type DebateRoundUpdateResponse,
   DebateWebSocketClient,
@@ -16,7 +17,8 @@ import {
 } from '@src/externals/websocket';
 import { useModal, useToast, useWebRTC } from '@src/hooks';
 import AiSummarizationModal from '@src/routes/Debate/_components/modal/AiSummarizationModal.tsx';
-import { useQueryClient } from '@tanstack/react-query';
+import SurveyModal from '@src/routes/Debate/_components/modal/SurveyModal';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useEffect, useEffectEvent, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
@@ -61,8 +63,18 @@ export const useDebateRealtimeConnection = (props: Props) => {
   const [isDebateJoined, setIsDebateJoined] = useState<boolean>(false);
   const [raisedHands, setRaisedHands] = useState<RaisedHandInfo[]>([]);
   const wsClientRef = useRef<DebateWebSocketClient | null>(null);
-  const { openModal } = useModal();
+  const { openModal, closeModal } = useModal();
   const { toast } = useToast();
+
+  const createSurveyMutation = useMutation({
+    mutationFn: createSurvey,
+    onSuccess: () => {
+      closeModal();
+    },
+    onError: (error) => {
+      toast.error(error instanceof Error ? error.message : '설문조사 제출에 실패했습니다.');
+    },
+  });
 
   // Voice 연결 상태
   const [voiceConnectionStatus, setVoiceConnectionStatus] =
@@ -255,18 +267,14 @@ export const useDebateRealtimeConnection = (props: Props) => {
   });
 
   /** 라운드 업데이트 */
-  const onDebateRoundUpdate = useEffectEvent(async (roundInfo: DebateRoundUpdateResponse) => {
+  const onDebateRoundUpdate = useEffectEvent((roundInfo: DebateRoundUpdateResponse) => {
     const roundType = roundInfo.payload.round.type as RoundType;
 
     if (roundType === 'PRESENTATION' || roundType === 'FREE') {
       onRoundStartBackdrop(roundType);
     }
 
-    if (roundType === 'PRESENTATION' && debateId) {
-      const debate = await queryClient.fetchQuery({
-        ...findOneDebateQueryOptions(debateId),
-        staleTime: 0,
-      });
+    if (roundType === 'PRESENTATION') {
       openModal(AiSummarizationModal, {
         bookTitle: `${debate.bookInfo.title} - ${debate.bookInfo.author}`,
         topic: debate.topic,
@@ -278,6 +286,12 @@ export const useDebateRealtimeConnection = (props: Props) => {
 
   /** 토론 종료 */
   const onDebateClosed = useEffectEvent(() => {
+    openModal(SurveyModal, {
+      onConfirm: (rate: number) => {
+        createSurveyMutation.mutate({ rate });
+      },
+      isLoading: createSurveyMutation.isPending,
+    });
     navigate('/home');
   });
